+++ packages/preset-react/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.7](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.3.6...@umijs/preset-react@1.3.7) (2020-03-05)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.3.6](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.3.5...@umijs/preset-react@1.3.6) (2020-03-05)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.3.5](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.3.4...@umijs/preset-react@1.3.5) (2020-03-05)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.3.4](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.3.3...@umijs/preset-react@1.3.4) (2020-03-05)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.3.3](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.3.2...@umijs/preset-react@1.3.3) (2020-03-04)
+
+**Note:** Version bump only for package @umijs/preset-react
+
+## [1.3.2](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.3.1...@umijs/preset-react@1.3.2) (2020-03-04)
+
+**Note:** Version bump only for package @umijs/preset-react
+
 ## [1.3.1](https://github.com/umijs/plugins/compare/@umijs/preset-react@1.3.0...@umijs/preset-react@1.3.1) (2020-03-02)
 
 **Note:** Version bump only for package @umijs/preset-react